changing project structure
