+++ book/src/document_history.md
@@ -1,4 +1,5 @@
 # Historical Completed Todo Lists
+
 This will be a useful reference when I want to put more effort into writing up
 this project. Additionally, it will help to track how much progress was made on
 what days, and give me a better idea of what kinds of goal-tracking works best
@@ -7,7 +8,9 @@ for my work flow.
 I should put more before the next sections, but I am starting this document a little late.
 
 ## Finishing up Item Coverage + Testing
+
 2025-03-09
+
 ```markdown
 ## Step 1: Expand Item Coverage ✓
 
@@ -23,17 +26,21 @@ Your code already handles several important Rust items (functions, structs, enum
    - Established `UseRelation` edges between items and their imports
    - Now tracking which external dependencies are being used
 
-3. **Type Aliases, Unions, and Trait Aliases**:
+3. **Type Aliases, Unions, and Trait Aliases**: ✓
    - Extend `TypeDefNode` enum to include these additional type definitions
    - These provide important aliasing and type relationship information
 
-4. **Constants and Statics**:
+4. **Constants and Statics**: ✓
    - Add `ValueNode` to represent constants and static variables
    - Track type information and initialization expressions
    - Important for understanding program constants and global state
 
-5. **Macros and Macro Rules**:
+5. **Macros and Macro Rules**: ✓
    - Create `MacroNode` to capture macro definitions
    - Record macro invocations as `MacroUseRelation`
    - This is critical for understanding Rust's meta-programming features
 ```
+
+## Make project more modular
+
+2025-03-10
