+++ src/error.rs
@@ -1,0 +1,2 @@
+// Custom Error Type and Handling
+// todo
