+++ src/lib.rs
@@ -1,1 +1,3 @@
 pub mod parser;
+pub mod serialization;
+// pub mod analysis; // For future code analysis features
