+++ src/parser/graph.rs
@@ -1,0 +1,30 @@
+use crate::parser::{
+    nodes::{FunctionNode, ImplNode, MacroNode, ModuleNode, TraitNode, TypeDefNode, ValueNode},
+    relations::Relation,
+    types::TypeNode,
+};
+
+use serde::{Deserialize, Serialize};
+
+// Main structure representing the entire code graph
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CodeGraph {
+    // Functions defined in the code
+    pub functions: Vec<FunctionNode>,
+    // Types (structs, enums) defined in the code
+    pub defined_types: Vec<TypeDefNode>,
+    // All observed types, including nested and generic types
+    pub type_graph: Vec<TypeNode>,
+    // Implementation blocks
+    pub impls: Vec<ImplNode>,
+    // Traits defined in the code
+    pub traits: Vec<TraitNode>,
+    // Relations between nodes
+    pub relations: Vec<Relation>,
+    // Modules defined in the code
+    pub modules: Vec<ModuleNode>,
+    // Constants and static variables
+    pub values: Vec<ValueNode>,
+    // Macros defined in the code
+    pub macros: Vec<MacroNode>,
+}
