+++ src/parser/types.rs
@@ -1,0 +1,132 @@
+use crate::parser::nodes::NodeId;
+
+use quote::ToTokens;
+use ron::ser::{to_string_pretty, PrettyConfig};
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::fs::File as FsFile;
+use std::io::Write;
+use std::path::Path;
+use syn::parse::Parser;
+use syn::ItemMod;
+use syn::{
+    visit::{self, Visit},
+    AngleBracketedGenericArguments, FnArg, GenericArgument, Generics, ItemEnum, ItemFn, ItemImpl,
+    ItemStruct, ItemTrait, Pat, PatIdent, PatType, PathArguments, ReturnType, Type, TypeParam,
+    TypePath, TypeReference, Visibility,
+};
+
+pub type TypeId = usize;
+
+// ANCHOR: TypeNode
+// Represents a type reference with full metadata
+#[derive(Debug, Serialize, Deserialize)]
+pub struct TypeNode {
+    pub id: TypeId,
+    pub kind: TypeKind,
+    // Reference to related types (e.g., generic arguments)
+    pub related_types: Vec<TypeId>,
+}
+//ANCHOR_END: TypeNode
+
+// ANCHOR: TypeKind_defn
+// Different kinds of types
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub enum TypeKind {
+    //ANCHOR_END: TypeKind_defn
+    Named {
+        path: Vec<String>, // Full path segments
+        is_fully_qualified: bool,
+    },
+    Reference {
+        lifetime: Option<String>,
+        is_mutable: bool,
+        // Type being referenced is in related_types[0]
+    },
+    Slice {
+        // Element type is in related_types[0]
+    },
+    Array {
+        // Element type is in related_types[0]
+        size: Option<String>,
+    },
+    Tuple {
+        // Element types are in related_types
+    },
+    // ANCHOR: ExternCrate
+    Function {
+        // Parameter types are in related_types (except last one)
+        // Return type is in related_types[last]
+        is_unsafe: bool,
+        is_extern: bool,
+        abi: Option<String>,
+    },
+    //ANCHOR_END: ExternCrate
+    Never,
+    Inferred,
+    RawPointer {
+        is_mutable: bool,
+        // Pointee type is in related_types[0]
+    },
+    // ANCHOR: TraitObject
+    TraitObject {
+        // Trait bounds are in related_types
+        dyn_token: bool,
+    },
+    //ANCHOR_END: TraitObject
+    // ANCHOR: ImplTrait
+    ImplTrait {
+        // Trait bounds are in related_types
+    },
+    //ANCHOR_END: ImplTrait
+    Paren {
+        // Inner type is in related_types[0]
+    },
+    // ANCHOR: ItemMacro
+    Macro {
+        name: String,
+        tokens: String,
+    },
+    //ANCHOR_END: ItemMacro
+    Unknown {
+        type_str: String,
+    },
+}
+
+// Represents a generic parameter
+#[derive(Debug, Serialize, Deserialize)]
+pub struct GenericParamNode {
+    pub id: NodeId,
+    pub kind: GenericParamKind,
+}
+
+// ANCHOR: generic_param_kind
+// Different kinds of generic parameters
+#[derive(Debug, Serialize, Deserialize)]
+pub enum GenericParamKind {
+    Type {
+        name: String,
+        bounds: Vec<TypeId>,
+        default: Option<TypeId>,
+    },
+    Lifetime {
+        name: String,
+        bounds: Vec<String>,
+    },
+    Const {
+        name: String,
+        type_id: TypeId,
+    },
+}
+//ANCHOR_END: generic_param_kind
+
+// Different kinds of visibility
+#[derive(Debug, Serialize, Deserialize, PartialEq, Eq)]
+pub enum VisibilityKind {
+    Public,
+    Crate,
+    Restricted(Vec<String>), // Path components of restricted visibility
+    Inherited,               // Default visibility
+}
+
+impl VisibilityKind {}
